Add ConfirmDialog
Add at interactable to indicate that the time waited to interactable should be set to the current waiting time
Add ConfirmDialog that confirms an event. In this commit, the "At Interactable" dialog was impletemented. This confirms that user is at an interactable part of a ride and how long they have been waiting.
Remove check for interactable from validRideEvent as a user may miss the interactable